+++ include/cec/CECEvent.h
@@ -1,0 +1,8 @@
+#pragma once
+
+enum class CECEvent
+{
+        On,
+        Off
+};
+