+++ libsrc/api/JsonAPI.cpp
@@ -868,12 +868,16 @@ void JsonAPI::handleConfigCommand(const QJsonObject &message, const QString &com
 	{
 		if (_adminAuthorized)
 		{
-			_hyperion->freeObjects(true);
+			Debug(_log, "Restarting due to RPC command");
+
 			Process::restartHyperion();
-			sendErrorReply("failed to restart hyperion", full_command, tan);
+
+			sendSuccessReply(command + "-" + subcommand, tan);
 		}
 		else
+		{
 			sendErrorReply("No Authorization", command, tan);
+		}
 	}
 	else
 	{
@@ -1409,7 +1413,6 @@ void JsonAPI::handleLedDeviceCommand(const QJsonObject &message, const QString &
 */	{
 		if (subc == "discover")
 		{
-
 			QJsonObject config;
 			config.insert("type", devType);