+++ libsrc/grabber/v4l2/V4L2Wrapper.cpp
@@ -90,7 +90,22 @@ bool V4L2Wrapper::getSignalDetectionEnable()
 	return _grabber.getSignalDetectionEnabled();
 }
 
+void V4L2Wrapper::setCecDetectionEnable(bool enable)
+{
+	_grabber.setCecDetectionEnable(enable);
+}
+
+bool V4L2Wrapper::getCecDetectionEnable()
+{
+	return _grabber.getCecDetectionEnabled();
+}
+
 void V4L2Wrapper::setDeviceVideoStandard(QString device, VideoStandard videoStandard)
 {
 	_grabber.setDeviceVideoStandard(device, videoStandard);
+}
+
+void V4L2Wrapper::handleCecEvent(CECEvent event)
+{
+	_grabber.handleCecEvent(event);
 }